# packages/specs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: App API
  version: 0.3.0
  description: |
    Spec-first HTTP API. Paths include the versioned prefix (`/api/v1`) so
    generated clients can be used without additional base-URL rewriting.

    **Authentication endpoints** live under `/api/v1/auth/*` and follow the
    [Better Auth protocol](https://www.better-auth.com/docs). They are
    deliberately not enumerated here: Better Auth owns that wire contract and
    evolves it independently of our OpenAPI surface. The `cookieAuth` and
    `bearerAuth` security schemes documented below describe how the *rest* of
    this API accepts credentials issued by that namespace — session cookies
    (web) and bearer tokens (mobile, via the `bearer` plugin).
  contact:
    name: App Team
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.example.com
    description: Production (placeholder)
...
paths:
...
  /api/v1/items:
    get:
      operationId: listItems
      summary: List saved items with optional filtering
      description: Returns a cursor-paginated list of the authenticated user's saved items. Supports filtering by read state, tag, favourite flag, and full-text search.
      tags: [items]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: readState
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ReadState"
          description: Filter by read state
        - name: tag
          in: query
          required: false
          schema:
            type: string
          description: Filter by tag slug
        - name: favorite
          in: query
          required: false
          schema:
            type: boolean
          description: Filter to favourites only
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Full-text search query
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque cursor for the next page
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Number of items to return (1–100, default 20)
      responses:
        "200":
          description: Paginated list of items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemList"
              examples:
                page1:
                  value:
                    items:
                      - id: clxabc123
                        url: https://example.com/article
                        canonicalUrl: null
                        title: "An interesting article"
                        siteName: "Example"
                        excerpt: "A short excerpt from the article."
                        leadImageUrl: null
                        faviconUrl: null
                        status: ready
                        extractStatus: ready
                        readState: unread
                        favorite: false
                        savedAt: "2026-06-13T10:00:00Z"
                        readAt: null
                        tags: ["tech", "reading"]
                    nextCursor: null
        "401":
          description: Not authenticated
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
                    detail: No active session

    post:
      operationId: saveItem
      summary: Save a new URL to the reading list
      description: Creates an item with status `pending`. A background job is dispatched to fetch metadata (title, excerpt, images). The item becomes `ready` or `failed` asynchronously.
      tags: [items]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveItemRequest"
            examples:
              basic:
                value:
                  url: https://example.com/article
                  tags: ["tech"]
      responses:
        "201":
          description: Item created (metadata extraction is asynchronous; initial status is `pending`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
              examples:
                created:
                  value:
                    id: clxabc123
                    url: https://example.com/article
                    canonicalUrl: null
                    title: null
                    siteName: null
                    excerpt: null
                    leadImageUrl: null
                    faviconUrl: null
                    status: pending
                    extractStatus: none
                    readState: unread
                    favorite: false
                    savedAt: "2026-06-13T10:00:00Z"
                    readAt: null
                    tags: ["tech"]
        "400":
          description: Request body invalid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                badBody:
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: "url must be a valid URI"
        "401":
          description: Not authenticated
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
                    detail: No active session
  /api/v1/items/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Item ID (cuid)

    get:
      operationId: getItem
      summary: Fetch a single item by ID
      description: Returns the full item record for the given ID, scoped to the authenticated user.
      tags: [items]
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: The requested item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
              examples:
                found:
                  value:
                    id: clxabc123
                    url: https://example.com/article
                    canonicalUrl: https://example.com/article
                    title: "An interesting article"
                    siteName: "Example"
                    excerpt: "A short excerpt."
                    leadImageUrl: https://example.com/img.jpg
                    faviconUrl: https://example.com/favicon.ico
                    status: ready
                    extractStatus: ready
                    readState: unread
                    favorite: false
                    savedAt: "2026-06-13T10:00:00Z"
                    readAt: null
                    tags: ["tech"]
        "401":
          description: Not authenticated
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
                    detail: No active session
        "404":
          description: Item not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: "Item clxabc123 does not exist"
                    code: item-not-found

    patch:
      operationId: updateItem
      summary: Update read state or favourite flag on an item
      description: Partially updates an item. At least one of `readState` or `favorite` must be provided. Setting `readState` to `read` for the first time records `readAt`.
      tags: [items]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemRequest"
            examples:
              markRead:
                value:
                  readState: read
              favourite:
                value:
                  favorite: true
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
              examples:
                updated:
                  value:
                    id: clxabc123
                    url: https://example.com/article
                    canonicalUrl: https://example.com/article
                    title: "An interesting article"
                    siteName: "Example"
                    excerpt: "A short excerpt."
                    leadImageUrl: null
                    faviconUrl: null
                    status: ready
                    extractStatus: ready
                    readState: read
                    favorite: false
                    savedAt: "2026-06-13T10:00:00Z"
                    readAt: "2026-06-13T11:00:00Z"
                    tags: ["tech"]
        "400":
          description: Request body invalid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                badBody:
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: "readState must be one of: unread, read, archived"
        "401":
          description: Not authenticated
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
                    detail: No active session
        "404":
          description: Item not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: "Item clxabc123 does not exist"
                    code: item-not-found

    delete:
      operationId: deleteItem
      summary: Permanently delete an item
      description: Permanently removes the item and all its tag associations. This operation cannot be undone.
      tags: [items]
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "204":
          description: Item deleted successfully
        "401":
          description: Not authenticated
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
                    detail: No active session
        "404":
          description: Item not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: "Item clxabc123 does not exist"
                    code: item-not-found
...
  /api/v1/items/{id}/highlights:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Item ID (cuid)

    get:
      operationId: listHighlights
      summary: List all highlights on an item
      description: Returns all highlights the authenticated user has created on the given item.
      tags: [highlights]
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: List of highlights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HighlightList"
              examples:
                list:
                  value:
                    highlights:
                      - id: clxhl001
                        itemId: clxabc123
                        quote: "Full article body in plain text."
                        prefix: "Returns the "
                        suffix: " for the given item."
                        note: null
                        color: yellow
                        createdAt: "2026-06-14T09:00:00Z"
        "401":
          description: Not authenticated
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
                    detail: No active session
        "404":
          description: Item not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: "Item clxabc123 does not exist"
                    code: item-not-found

    post:
      operationId: createHighlight
      summary: Create a highlight on an item
      description: Saves a new text highlight (with optional note and color) on the given item.
      tags: [highlights]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHighlightRequest"
            examples:
              basic:
                value:
                  quote: "Full article body in plain text."
                  prefix: "Returns the "
                  suffix: " for the given item."
                  color: yellow
              withNote:
                value:
                  quote: "Full article body in plain text."
                  prefix: "Returns the "
                  suffix: " for the given item."
                  note: "Important paragraph"
                  color: green
      responses:
        "201":
          description: Highlight created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Highlight"
              examples:
                created:
                  value:
                    id: clxhl001
                    itemId: clxabc123
                    quote: "Full article body in plain text."
                    prefix: "Returns the "
                    suffix: " for the given item."
                    note: null
                    color: yellow
                    createdAt: "2026-06-14T09:00:00Z"
        "400":
          description: Request body invalid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                badBody:
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: "quote is required"
        "401":
          description: Not authenticated
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
                    detail: No active session
        "404":
          description: Item not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: "Item clxabc123 does not exist"
                    code: item-not-found
...
  /api/v1/highlights/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Highlight ID (cuid)

    patch:
      operationId: updateHighlight
      summary: Update a highlight's note or color
      description: Partially updates a highlight. At least one of `note` or `color` must be provided.
      tags: [highlights]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHighlightRequest"
            examples:
              changeColor:
                value:
                  color: green
              addNote:
                value:
                  note: "Key insight here"
      responses:
        "200":
          description: Updated highlight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Highlight"
              examples:
                updated:
                  value:
                    id: clxhl001
                    itemId: clxabc123
                    quote: "Full article body in plain text."
                    prefix: "Returns the "
                    suffix: " for the given item."
                    note: "Key insight here"
                    color: green
                    createdAt: "2026-06-14T09:00:00Z"
        "400":
          description: Request body invalid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                badBody:
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: "At least one of note or color must be provided"
        "401":
          description: Not authenticated
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
                    detail: No active session
        "404":
          description: Highlight not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: "Highlight clxhl001 does not exist"
                    code: highlight-not-found

    delete:
      operationId: deleteHighlight
      summary: Delete a highlight
      description: Permanently removes the highlight. This operation cannot be undone.
      tags: [highlights]
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "204":
          description: Highlight deleted successfully
        "401":
          description: Not authenticated
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
                    detail: No active session
        "404":
          description: Highlight not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: "Highlight clxhl001 does not exist"
                    code: highlight-not-found
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Better Auth bearer token
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details
      required: [type, title, status]
      properties:
        type:
          type: ["string", "null"]
          format: uri
          default: about:blank
        title:
          type: ["string", "null"]
        status:
          type: integer
          format: int32
        detail:
          type: ["string", "null"]
        instance:
          type: ["string", "null"]
        code:
          type: ["string", "null"]
    ItemStatus:
      type: string
      enum: [pending, ready, failed]
      description: Processing status of the item's metadata extraction job
    ReadState:
      type: string
      enum: [unread, read, archived]
      description: The user's read progress for an item
    Item:
      type: object
      required: [id, url, status, extractStatus, readState, favorite, savedAt, tags]
      properties:
        id:
          type: string
          description: Unique item ID (cuid)
        url:
          type: string
          format: uri
          description: Original URL submitted by the user
        canonicalUrl:
          type: ["string", "null"]
          format: uri
          description: Canonical URL resolved during metadata extraction
        title:
          type: ["string", "null"]
          description: Page title
        siteName:
          type: ["string", "null"]
          description: Name of the publishing site
        excerpt:
          type: ["string", "null"]
          description: Short plain-text excerpt
        leadImageUrl:
          type: ["string", "null"]
          format: uri
          description: Hero/lead image URL
        faviconUrl:
          type: ["string", "null"]
          format: uri
          description: Site favicon URL
        status:
          $ref: "#/components/schemas/ItemStatus"
        extractStatus:
          $ref: "#/components/schemas/ExtractStatus"
        readState:
          $ref: "#/components/schemas/ReadState"
        favorite:
          type: boolean
          description: Whether the item is marked as a favourite
        savedAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp when the item was saved
        readAt:
          type: ["string", "null"]
          format: date-time
          description: ISO-8601 timestamp when the item was first marked read
        tags:
          type: array
          items:
            type: string
          description: Slugs of tags attached to this item
    ItemList:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        nextCursor:
          type: ["string", "null"]
          description: Opaque cursor for the next page; null when no more pages
...
    SaveItemRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: URL to save
        tags:
          type: array
          items:
            type: string
          description: Optional tag slugs to attach immediately
    UpdateItemRequest:
      type: object
      minProperties: 1
      description: At least one of `readState` or `favorite` must be provided.
      properties:
        readState:
          $ref: "#/components/schemas/ReadState"
        favorite:
          type: boolean
          description: Set or clear the favourite flag
...
    ExtractStatus:
      type: string
      enum: [none, extracting, ready, failed]
      description: Status of the article content extraction job for an item
...
    HighlightColor:
      type: string
      enum: [yellow, green, blue, pink]
      description: Color label for a text highlight
    Highlight:
      type: object
      required: [id, itemId, quote, prefix, suffix, color, createdAt]
      properties:
        id:
          type: string
          description: Unique highlight ID (cuid)
        itemId:
          type: string
          description: ID of the item this highlight belongs to (cuid)
        quote:
          type: string
          description: The exact text that was highlighted
        prefix:
          type: string
          description: Short text immediately before the highlighted quote (for anchor context)
        suffix:
          type: string
          description: Short text immediately after the highlighted quote (for anchor context)
        note:
          type: ["string", "null"]
          description: Optional reader note attached to the highlight
        color:
          $ref: "#/components/schemas/HighlightColor"
        createdAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp when the highlight was created
    HighlightList:
      type: object
      required: [highlights]
      properties:
        highlights:
          type: array
          items:
            $ref: "#/components/schemas/Highlight"
    CreateHighlightRequest:
      type: object
      required: [quote]
      properties:
        quote:
          type: string
          description: The exact text that was highlighted
        prefix:
          type: string
          description: Short text immediately before the quote (for anchor context)
        suffix:
          type: string
          description: Short text immediately after the quote (for anchor context)
        note:
          type: string
          description: |
            Optional reader note to attach. To clear a note on an existing
            highlight, use `PATCH /highlights/{id}` with `note: null` —
            create only sets a note, it cannot null one.
        color:
          $ref: "#/components/schemas/HighlightColor"
    UpdateHighlightRequest:
      type: object
      minProperties: 1
      description: At least one of `note` or `color` must be provided.
      properties:
        note:
          type: ["string", "null"]
          description: Updated reader note; set to null to clear
        color:
          $ref: "#/components/schemas/HighlightColor"